Revert "fixing user replies and liking"
