+++ client/src/api/queries/thread.ts
@@ -141,7 +141,6 @@ export const fetchThreadReplies = async (
   if (userId) {
     params.append("userId", userId);
   }
-
   if (limit !== undefined) {
     params.append("limit", limit.toString());
   }
