+++ server/storage.ts
@@ -1195,8 +1195,8 @@ export class DatabaseStorage implements IStorage {
         `getRepliesByThread: threadId=${threadId}, limit=${limit}, offset=${offset}`,
       );
 
-      // Build query with explicit column selection and apply pagination
-      const threadReplies = await db
+      // Build query with explicit column selection
+      const baseQuery = db
         .select({
           id: replies.id,
           threadId: replies.threadId,
@@ -1210,10 +1210,20 @@ export class DatabaseStorage implements IStorage {
         })
         .from(replies)
         .where(eq(replies.threadId, threadId))
-        .orderBy(replies.createdAt)
-        .limit(limit || 1000) // Use a reasonable default if limit is undefined
-        .offset(offset || 0); // Use 0 as default if offset is undefined
+        .orderBy(replies.createdAt);
 
+      // Apply pagination
+      let query = baseQuery;
+      if (limit !== undefined) {
+        query = query.limit(limit);
+        console.log(`Applied limit: ${limit}`);
+      }
+      if (offset !== undefined) {
+        query = query.offset(offset);
+        console.log(`Applied offset: ${offset}`);
+      }
+
+      const threadReplies = await query;
       console.log(`Database query returned ${threadReplies.length} replies`);
 
       // If we have a current user, check which replies they've liked
@@ -1265,7 +1275,7 @@ export class DatabaseStorage implements IStorage {
         console.log("Attempting to reconnect to database...");
         try {
           // Use the existing db instance instead of creating a new one
-          const threadReplies = await db
+          const baseQuery = db
             .select({
               id: replies.id,
               threadId: replies.threadId,
@@ -1279,9 +1289,18 @@ export class DatabaseStorage implements IStorage {
             })
             .from(replies)
             .where(eq(replies.threadId, threadId))
-            .orderBy(replies.createdAt)
-            .limit(limit || 1000) // Use a reasonable default if limit is undefined
-            .offset(offset || 0); // Use 0 as default if offset is undefined
+            .orderBy(replies.createdAt);
+
+          // Apply pagination
+          let query = baseQuery;
+          if (limit !== undefined) {
+            query = query.limit(limit);
+          }
+          if (offset !== undefined) {
+            query = query.offset(offset);
+          }
+
+          const threadReplies = await query;
 
           // If we have a current user, check which replies they've liked
           if (currentUserId) {
